archive: fix chmod fallback for device nodes on nodev mounts
Commit [f43ca4d][1] introduced an O_RDONLY-based chmod fallback for systems where
fchmodat with AT_SYMLINK_NOFOLLOW is unsupported. Opening a device node this
way fails with EACCES when the filesystem is mounted with nodev.

Use O_PATH to reference the entry without opening the underlying device, then
apply the mode through /proc/self/fd. Keep the existing fallback on non-Linux
Unix platforms.

[1]: https://github.com/moby/go-archive/commit/f43ca4dfa3054c1ff02295d5f40a7cc391225ae4
