+++ archive_unix.go
@@ -129,23 +129,6 @@ func chmodNoSymlink(root *os.Root, name string, mode os.FileMode) error {
 	return chmodNoSymlinkFallback(int(parent.Fd()), base, name, perm) // #nosec G115 -- ignore integer overflow conversion for parent.Fd
 }
 
-// chmodNoSymlinkFallback applies mode without following the final path
-// component on systems without fchmodat2 support.
-//
-// Callers must have already excluded symlink entries.
-func chmodNoSymlinkFallback(parentFD int, base, name string, perm uint32) error {
-	fd, err := unix.Openat(parentFD, base, unix.O_RDONLY|unix.O_NOFOLLOW|unix.O_NONBLOCK, 0)
-	if err != nil {
-		return &os.PathError{Op: "openat", Path: name, Err: err}
-	}
-	defer unix.Close(fd)
-
-	if err := unix.Fchmod(fd, perm); err != nil {
-		return &os.PathError{Op: "fchmod", Path: name, Err: err}
-	}
-	return nil
-}
-
 // fileModeToPerm returns the subset of an os.FileMode that can be applied
 // by chmod.
 func fileModeToPerm(mode os.FileMode) uint32 {
